updated cro to run at 13:30 IST

+++ src/dst/pgsql.py
@@ -75,11 +75,12 @@ class PgSQLSaver:
         if(exists):
             ## Drop indexes if exist (except primary key index)
             for index_name, index_def in indexes.items():
-                with conn.cursor() as curs:
-                    sql_stmt = f"DROP INDEX IF EXISTS \"{index_name}\";"
-                    self.inform(sql_stmt)
-                    curs.execute(sql_stmt)
-                    conn.commit()
+                if "_pkey" not in index_name.lower():
+                    with conn.cursor() as curs:
+                        sql_stmt = f"DROP INDEX IF EXISTS \"{index_name}\";"
+                        self.inform(sql_stmt)
+                        curs.execute(sql_stmt)
+                        conn.commit()
             self.inform("Dropped indexes")
         conn.close()
         return exists
